Move ssh code into cloudprovider/common package
